# Speed-up embedding composites for subgraph embeddings