Embedding step is ~100x faster and unembedding is ~20x faster (for subgraphs).

(On my laptop, for a Z12 graph, that's 1.7s vs 17ms for embedding and 170ms vs 9ms for unembedding.)

Close #579.